requirements.txt
pillow==9.2.0
numpy==1.24.0